+++ modules/orca/main.tf
@@ -1,9 +1,3 @@
-## Local Variables
-locals {
-  tags = merge(var.tags, { Deployment = var.prefix })
-}
-
-
 ## Referenced Modules
 
 ## orca_lambdas - lambdas module that calls iam and security_groups module
@@ -21,7 +15,7 @@ module "orca_lambdas" {
   rds_security_group_id                                = var.rds_security_group_id
   vpc_id                                               = var.vpc_id
   ## OPTIONAL
-  tags                           = local.tags
+  tags                           = var.tags
   default_multipart_chunksize_mb = var.default_multipart_chunksize_mb
 
   ## --------------------------
@@ -67,7 +61,7 @@ module "orca_lambdas_secondary" {
   lambda_subnet_ids                                    = var.lambda_subnet_ids
   prefix                                               = var.prefix
   ## OPTIONAL
-  tags                           = local.tags
+  tags                           = var.tags
   default_multipart_chunksize_mb = var.default_multipart_chunksize_mb
 
   ## --------------------------
@@ -78,11 +72,13 @@ module "orca_lambdas_secondary" {
   orca_sqs_internal_report_queue_id             = module.orca_sqs.orca_sqs_internal_report_queue_id
   orca_sqs_s3_inventory_queue_arn               = module.orca_sqs.orca_sqs_s3_inventory_queue_arn
   restore_object_role_arn                       = module.orca_iam.restore_object_role_arn
+  orca_reports_bucket_arn                       = var.orca_reports_bucket_arn
   vpc_postgres_ingress_all_egress_id            = module.orca_lambdas.vpc_postgres_ingress_all_egress_id
 
   ## OPTIONAL
   orca_reconciliation_lambda_memory_size = var.orca_reconciliation_lambda_memory_size
   orca_reconciliation_lambda_timeout     = var.orca_reconciliation_lambda_timeout
+  s3_report_frequency                    = var.s3_report_frequency
 }
 
 
@@ -99,7 +95,7 @@ module "orca_workflows" {
   workflow_config = var.workflow_config
 
   ## OPTIONAL
-  tags = local.tags
+  tags = var.tags
 
   ## --------------------------
   ## ORCA Variables
@@ -126,7 +122,7 @@ module "orca_iam" {
   permissions_boundary_arn = var.permissions_boundary_arn
   prefix                   = var.prefix
   # OPTIONAL
-  tags = local.tags
+  tags = var.tags
   # --------------------------
   # ORCA Variables
   # --------------------------
@@ -147,7 +143,7 @@ module "orca_secretsmanager" {
   prefix = var.prefix
 
   ## OPTIONAL
-  tags = local.tags
+  tags = var.tags
   ## --------------------------
   ## ORCA Variables
   ## --------------------------
@@ -175,7 +171,7 @@ module "orca_sqs" {
   prefix = var.prefix
 
   ## OPTIONAL
-  tags = local.tags
+  tags = var.tags
 
   ## --------------------------
   ## ORCA Variables