+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and includes an additional section for migration notes.
     Adds `orca_reconciliation_lambda_memory_size` and `orca_reconciliation_lambda_timeout` to Terraform variables.
 - *ORCA-308* Added lambda perform_orca_reconcile to find differences between S3 Inventory reports and Orca catalog.
 - *ORCA-403* Added lambda post_to_queue_and_trigger_step_function to trigger step function for internal reconciliation.
+- *ORCA-373* Added input variable for `orca_reports_bucket_arn`. Set in your `variables.tf` or `orca_variables.tf` file as shown below.
+    Report frequency defaults to `Daily`, but can be set to `Weekly` through variable `s3_report_frequency`.
 
 ### Changed
 - *ORCA-299* `db_deploy` task has been updated to deploy ORCA internal reconciliation tables and objects.
@@ -32,8 +34,23 @@ and includes an additional section for migration notes.
 
 ### Migration Notes
 
+- Create a new bucket `PREFIX-orca-reports` in the same account and region as your other buckets.
+  - Give the bucket a lifecycle configuration with an expiration period of 30 days.
+  - TODO: Add cross account policy similar to https://nasa.github.io/cumulus-orca/docs/developer/deployment-guide/deployment-s3-bucket/
+  - Update the `buckets` variable in your `tfvars` file by adding the bucket.
+    ```
+    buckets = {
+      ...
+      orca_reports = {
+        name = "PREFIX-orca-reports"
+        type = "orca-reports"
+      },
+      ...
+    }
+    ```
 - The user should update their `orca.tf`, `variables.tf` and `terraform.tfvars` files with new variables. The following required variables have been added:
   - dlq_subscription_email
+  - orca_reports_bucket_arn
   - s3_access_key
   - s3_secret_key
   
@@ -43,6 +60,11 @@ and includes an additional section for migration notes.
 variable "dlq_subscription_email" {
   type        = string
   description = "The email to notify users when messages are received in dead letter SQS queue due to restore failure. Sends one email until the dead letter queue is emptied."
+}
+
+variable "orca_reports_bucket_arn" {
+  type        = string
+  description = "The ARN of the bucket to store s3 inventory reports."
 }
 
 variable "s3_access_key" {
@@ -80,11 +102,12 @@ variable "s3_secret_key" {
   ## ORCA Variables
   ## --------------------------
   ## REQUIRED
-  orca_default_bucket     = var.orca_default_bucket
   db_admin_password       = var.db_admin_password
   db_user_password        = var.db_user_password
   db_host_endpoint        = var.db_host_endpoint
   dlq_subscription_email  = var.dlq_subscription_email
+  orca_default_bucket     = var.orca_default_bucket
+  orca_reports_bucket_arn = var.orca_reports_bucket_arn
   rds_security_group_id   = var.rds_security_group_id
   s3_access_key           = var.s3_access_key
   s3_secret_key           = var.s3_secret_key
@@ -104,6 +127,7 @@ variable "s3_secret_key" {
   orca_recovery_retry_interval                         = 1
   orca_recovery_retry_backoff                          = 2
   s3_inventory_queue_message_retention_time_seconds    = 432000
+  s3_report_frequency                                  = "Daily"
   sqs_delay_time_seconds                               = 0
   sqs_maximum_message_size                             = 262144
   staged_recovery_queue_message_retention_time_seconds = 432000